requirements.txt
transformers
discord.py
python-dotenv
gpt4all
transformers[sentencepiece]